업로드 함수 리팩토링 #167
