+++ app/Views/csl/bulk/write.php
@@ -34,7 +34,7 @@
                                 <div class="form-group row">
                                     <label for="bulk_file" class="col-sm-2 col-form-label">엑셀파일</label>
                                     <div class="col-sm-5">
-                                        <input type="file" class="form-control" id="bulk_file" name="bulk_file" onchange="upload(this.id)">
+                                        <input type="file" class="form-control" id="bulk_file" name="bulk_file" onchange="upload(this.id, 'general')">
                                     </div>
                                     <div class="col-sm-5">
                                         <ul class="list-unstyled" id="bulk_file_visible"></ul>
@@ -74,11 +74,12 @@
     });
 
     function upload_after(proc_result) {
-        var file_id = proc_result.file_id;
-        var input_file_id = proc_result.input_file_id;
-        var file_name_org = proc_result.file_name_org;
+        var info = proc_result.info;
+        var file_id = info.file_id;
+        var input_file_id = info.input_file_id;
+        var file_name_org = info.file_name_org;
 
         $("#"+input_file_id+"_hidden").val(file_id);
-        $("#"+input_file_id+"_visible").html("<li id='"+file_id+"'><a href='/file/download/"+file_id+"'>"+file_name_org+"</a><span class='ml-3'><button type='button' id='"+file_id+"' class='btn btn-danger btn-xs' onclick='file_delete(\""+file_id+"\")'>삭제</button></span></li>");
+        $("#"+input_file_id+"_visible").html("<li id='"+file_id+"'><a href='/download/download/"+file_id+"'>"+file_name_org+"</a><span class='ml-3'><button type='button' id='"+file_id+"' class='btn btn-danger btn-xs' onclick='file_delete(\""+file_id+"\")'>삭제</button></span></li>");
     }
 </script>
